+++ src/phys/time/delta/ops.rs
@@ -7,9 +7,7 @@
 // - impl_ops
 
 use super::*;
-use crate::unwrap;
-#[allow(unused)]
-use crate::{Float, FloatExt};
+use crate::{Float, unwrap};
 
 /// # Operations.
 impl TimeDelta {
@@ -161,8 +159,8 @@ impl TimeDelta {
     ///
     /// # Panics
     /// Panics if the given float overflows the minimum or maximum time delta values.
-    pub fn from_secs_f64(secs: f64) -> TimeDelta {
-        TimeDelta::try_from_secs_f64(secs).expect("finite and in-bounds f64")
+    pub const fn from_secs_f64(secs: f64) -> TimeDelta {
+        unwrap![ok_expect TimeDelta::try_from_secs_f64(secs), "finite and in-bounds f64"]
     }
     /// Returns a time delta corresponding to the number of seconds.
     ///
@@ -170,8 +168,8 @@ impl TimeDelta {
     ///
     /// # Panics
     /// Panics if the given float overflows the minimum or maximum time delta values.
-    pub fn from_secs_f32(secs: f32) -> TimeDelta {
-        TimeDelta::try_from_secs_f32(secs).expect("finite and in-bounds f32")
+    pub const fn from_secs_f32(secs: f32) -> TimeDelta {
+        unwrap![ok_expect TimeDelta::try_from_secs_f32(secs), "finite and in-bounds f32"]
     }
 
     /// Returns a time delta corresponding to the number of seconds.
@@ -180,15 +178,8 @@ impl TimeDelta {
     ///
     /// If the given float overflows the minimum or maximum time delta values,
     /// then an error is returned.
-    //
-    // TODO: RETHINK
-    // # Features
-    // - Uses `std` or `_float_f64` when available, leveraging
-    //   `trunc`, `fract`, and `round` for precise, bias-free conversion.
-    // - In strict `no_std` mode, manually rounds using integer arithmetic,
-    //   ensuring correctness while lacking exact fractional nanosecond precision.
     #[rustfmt::skip]
-    pub fn try_from_secs_f64(secs: f64) -> Result<TimeDelta, &'static str> {
+    pub const fn try_from_secs_f64(secs: f64) -> Result<TimeDelta, &'static str> {
         if !secs.is_finite() {
             return Err("could not convert non-finite seconds {secs} to time delta"); }
         if secs < (i64::MIN as f64) {
@@ -196,16 +187,8 @@ impl TimeDelta {
         if secs > (i64::MAX as f64) {
             return Err("floating point seconds {secs} overflows TimeDelta::MAX"); }
         let (isecs, nanos);
-        // #[cfg(any(feature = "std", feature = "_float_f64"))] {
-        isecs = secs.trunc() as i64;
-        nanos = (secs.fract() * NANOS_PER_SEC as f64).round() as i32;
-        // }
-        // #[cfg(not(any(feature = "std", feature = "_float_f64")))] {
-        //     let secs_rounded = if secs >= 0.0 { secs + 0.5 }  // Round normally
-        //     else { secs - 0.5 };  // Round away from zero for negatives
-        //     isecs = secs_rounded as i64;
-        //     nanos = ((secs_rounded - isecs as f64) * NANOS_PER_SEC as f64) as i32;
-        // }
+        isecs = Float(secs).trunc().0 as i64;
+        nanos = Float(Float(secs).fract().0 * NANOS_PER_SEC as f64).round().0 as i32;
         Ok(TimeDelta::new_unchecked(isecs, nanos))
     }
 
@@ -213,10 +196,9 @@ impl TimeDelta {
     ///
     /// The number given may have a fractional nanosecond component.
     ///
-    /// If the given float overflows the minimum or maximum time delta
-    /// values, then an error is returned.
+    /// Returns an error if the given float overflows the minimum or maximum time delta values.
     #[rustfmt::skip]
-    pub fn try_from_secs_f32(secs: f32) -> Result<TimeDelta, &'static str> {
+    pub const fn try_from_secs_f32(secs: f32) -> Result<TimeDelta, &'static str> {
         if !secs.is_finite() {
             return Err("could not convert non-finite seconds {secs} to time delta"); }
         if secs < (i64::MIN as f32) {
@@ -224,14 +206,8 @@ impl TimeDelta {
         if secs > (i64::MAX as f32) {
             return Err("floating point seconds {secs} overflows TimeDelta::MAX"); }
         let (isecs, nanos);
-        isecs = secs.trunc() as i64;
-        nanos = (secs.fract() * NANOS_PER_SEC as f32).round() as i32;
-        // #[cfg(not(any(feature = "std", feature = "_float_f32")))] {
-        //     let secs_rounded = if secs >= 0.0 { secs + 0.5 }  // Round normally
-        //     else { secs - 0.5 };  // Round away from zero for negatives
-        //     isecs = secs_rounded as i64;
-        //     nanos = ((secs_rounded - isecs as f32) * NANOS_PER_SEC as f32) as i32;
-        // }
+        isecs = Float(secs).trunc().0 as i64;
+        nanos = Float(Float(secs).fract().0 * NANOS_PER_SEC as f32).round().0 as i32;
         Ok(TimeDelta::new_unchecked(isecs, nanos))
     }
 
@@ -252,29 +228,17 @@ impl TimeDelta {
     ///
     /// # Panics
     /// Panics if the given float overflows the minimum or maximum time delta values.
-    pub fn from_millis_f64(millis: f64) -> TimeDelta {
-        TimeDelta::try_from_millis_f64(millis).expect("finite and in-bounds f64")
-    }
-    /// Compile-time friendly version of `try_from_millis_f64`.
-    pub const fn const_from_millis_f64(millis: f64) -> TimeDelta {
-        unwrap![ok_expect TimeDelta::const_try_from_millis_f64(millis), "finite and in-bounds f64"]
+    pub const fn from_millis_f64(millis: f64) -> TimeDelta {
+        unwrap![ok_expect TimeDelta::try_from_millis_f64(millis), "finite and in-bounds f64"]
     }
 
     /// Returns a time delta corresponding to the number of milliseconds.
     ///
     /// The number given may have a fractional nanosecond component.
     ///
-    /// If the given float overflows the minimum or maximum time delta
-    /// values, then an error is returned.
-    //
-    // TODO: RETHINK
-    // # Features
-    // - Uses `std` or `_float_f64` when available, leveraging `round`, `div_euclid`,
-    //   and `rem_euclid` for precise, bias-free conversion.
-    // - In strict `no_std` mode, manually rounds using integer arithmetic, ensuring correctness
-    //   while lacking exact Euclidean division for negatives.
+    /// Returns an error if the given float overflows the minimum or maximum time delta values.
     #[rustfmt::skip]
-    pub fn try_from_millis_f64(millis: f64) -> Result<TimeDelta, &'static str> {
+    pub const fn try_from_millis_f64(millis: f64) -> Result<TimeDelta, &'static str> {
         if !millis.is_finite() {
             return Err("could not convert non-finite milliseconds {millis} to time delta"); }
         if millis < (i64::MIN as f64) {
@@ -282,52 +246,10 @@ impl TimeDelta {
         if millis > (i64::MAX as f64) {
             return Err("floating point milliseconds {millis} overflows TimeDelta::MAX"); }
         let (millis_rounded, secs, nanos);
-        // #[cfg(any(feature = "std", feature = "_float_f64"))]
-        // {
-        millis_rounded = millis.round();
-        secs = millis_rounded.div_euclid(1_000.0) as i64;
-        nanos = (millis_rounded.rem_euclid(1_000.0) * 1_000_000.0) as i32;
-        // }
-        // #[cfg(not(any(feature = "std", feature = "_float_f64")))]
-        // {
-        //     millis_rounded = if millis >= 0.0 { millis + 0.5 }  // Round normally
-        //     else { millis - 0.5 };  // Round away from zero for negatives
-        //     let millis_i64 = millis_rounded as i64;
-        //     secs = millis_i64 / MILLIS_PER_SEC;
-        //     nanos = ((millis_i64 % MILLIS_PER_SEC) * NANOS_PER_MILLI as i64) as i32;
-        // }
+        millis_rounded = Float(millis).round().0;
+        secs = Float(millis_rounded).div_euclid(1_000.0).0 as i64;
+        nanos = (Float(millis_rounded).rem_euclid(1_000.0).0 * 1_000_000.0) as i32;
         Ok(TimeDelta::new_unchecked(secs, nanos))
-    }
-    /// Compile-time friendly version of `try_from_millis_f64`.
-    //
-    // TODO: RETHINK
-    // # Features
-    // - Uses `_float_f64` if enabled, leveraging [`Float`]'s `const_round`, `div_euclid`,
-    //   and `rem_euclid` for precise, bias-free conversion.
-    // - Without `_float_f64`, rounds manually using integer arithmetic, preventing
-    //   systematic underestimation but lacking exact Euclidean division for negatives.
-    #[rustfmt::skip]
-    pub const fn const_try_from_millis_f64(millis: f64) -> Result<TimeDelta, &'static str> {
-        if !millis.is_finite() {
-            return Err("could not convert non-finite milliseconds {millis} to time delta"); }
-        if millis < (i64::MIN as f64) {
-            return Err("floating point milliseconds {millis} overflows TimeDelta::MIN"); }
-        if millis > (i64::MAX as f64) {
-            return Err("floating point milliseconds {millis} overflows TimeDelta::MAX"); }
-        let (millis_rounded, secs, nanos);
-        millis_rounded = Float(millis).const_round().0 as i64;
-        secs = Float(millis_rounded as f64).div_euclid(MILLIS_PER_SEC as f64).0 as i64;
-        nanos = Float(millis_rounded as f64).rem_euclid(MILLIS_PER_SEC as f64).0 as i64
-            * NANOS_PER_MILLI as i64;
-        // #[cfg(not(feature = "_float_f64"))]
-        // {
-        //     // millis_rounded = millis as i64; // slight systematic underestimation
-        //     millis_rounded = if millis >= 0.0 { (millis + 0.5) as i64 }  // Round normally
-        //     else { (millis - 0.5) as i64 };  // Round away from zero for negatives
-        //     secs = millis_rounded / MILLIS_PER_SEC;
-        //     nanos = (millis_rounded % MILLIS_PER_SEC) * NANOS_PER_MILLI as i64;
-        // }
-        Ok(TimeDelta::new_unchecked(secs, nanos as i32))
     }
 
     /* ops */
@@ -336,14 +258,14 @@ impl TimeDelta {
     ///
     /// # Panics
     /// This panics if the result is not finite or overflows a `TimeDelta`.
-    pub fn mul_f64(self, rhs: f64) -> TimeDelta {
+    pub const fn mul_f64(self, rhs: f64) -> TimeDelta {
         TimeDelta::from_secs_f64(rhs * self.as_secs_f64())
     }
     /// Returns the result of multiplying this duration by the given 32-bit float.
     ///
     /// # Panics
     /// This panics if the result is not finite or overflows a `TimeDelta`.
-    pub fn mul_f32(self, rhs: f32) -> TimeDelta {
+    pub const fn mul_f32(self, rhs: f32) -> TimeDelta {
         TimeDelta::from_secs_f32(rhs * self.as_secs_f32())
     }
 
@@ -351,14 +273,14 @@ impl TimeDelta {
     ///
     /// # Panics
     /// This panics if the result is not finite or overflows a `TimeDelta`.
-    pub fn div_f64(self, rhs: f64) -> TimeDelta {
+    pub const fn div_f64(self, rhs: f64) -> TimeDelta {
         TimeDelta::from_secs_f64(self.as_secs_f64() / rhs)
     }
     /// Returns the result of dividing this duration by the given `f32`.
     ///
     /// # Panics
     /// This panics if the result is not finite or overflows a `TimeDelta`.
-    pub fn div_f32(self, rhs: f32) -> TimeDelta {
+    pub const fn div_f32(self, rhs: f32) -> TimeDelta {
         TimeDelta::from_secs_f32(self.as_secs_f32() / rhs)
     }